reverting attempt to generate markdown
(wip) towards generating user guide

(wip) Figuring out how the website/guide is generated and (at the same time) write UserPageGuides containing validating snippets.

Spec5: the code blocks seems that are being evaluated both when it is ran as a Specification and when it is ran as a UserGuidePage.

issue: But, the HTML page renders badly!

issue: for some reason, some specs, Spec5 included, are not being evaluated from $ sbt test but are (being evaluated) if ran from IDE.